feat: Fixed up navigation issue in details fragment
Because we are using a ViewPager in the details fragment, the
up button of the currently displayed fragment was being covered
by adjacent fragments. There wasn't a simple solution to being
able to use the native up button for the currently displayed
fragment. Instead, the up button was disabled and a ImageButton
with an appropriate drawable was used to mimic the up button.

In addition, the default title within the CollapsingToolbarLayout
was removed and is now replaced by the logo for the app.

I started working on adjusting the toolbar for the list activity
as well, but then switched gears to straighten out the up button
issue instead. I was ultimately hoping to have the app bar within
an AppBarLayout + CollapsingToolbarLayout combo and then set the
elevation of the app bar per Material Design guidelines. This
will need to be taken care of at some point.