feat: Adjusted app bar to follow MD metrics
The margins of the app bar have been adjusted to follow the
metrics of the MD spec for both the main activity and the detail
activity/fragment.

There is a minor error with the app bar in the detail activity,
where the logo is partially hidden by the window insets. I will
need to investigate this some more in order to resolve this issue.