feat: Formatted body text when added to DB
Previously, the body text of the detail fragment was formatted
right before it was added to the TextView that would contain the
body text. What's done instead is the body text is properly
formatted immediately when it's added to the SQLiteDB. A
Utilities class was added which contains a sanitizeBodyText()
method.