feat: Adjusted list activity to a staggered grid layout
Previously, a regular grid layout was used for the list activity.
This was primarily done so that we could use a ItemDecoration on
the grid which allowed for equal spacing along each edge of the
child views. Initially, it wasn't thought this could be
accomplished when using a StaggeredGridLayout. However, upon
additional research, it was deemed to be possible. A new
StaggeredGridItemDecoration has been created which allows for
equal spacing for a staggered grid as well. This item decoration
is applied to the staggered grid and everything looks nice and
neatly prepared on screen :D