feat: Slight adjustment to transition name for shared elements
Changed the ID that is passed to the transition name of the shared
element transition. Previously, I was using the article ID.
However, I chose to use the adapter position of both the adapter
within the RecyclerView and the adapter of the ViewPager.
Although these are 2 different adapters, I discovered that the
position of the adapters always remains the same (likely because
the underlying data within the SQLite DB is sorted in the same
order).